Concurrency test

+++ src/ntapfuse_ops.c
@@ -61,6 +61,54 @@ fullpath (const char *path, char *buf)
   strcat (buf, path);
 }
 
+int get_lockfile(int file_descriptor) {
+
+  char lock_path[PATH_MAX];
+  fullpath("/lock", lock_path);
+
+  int count = 1;
+  int lock = 0;
+
+  // Determine if additional sleep should be done
+  // based on environment variables
+  int sleep_time = 0;
+  const char * st = getenv("SLEEP_TIME");
+  if (st != NULL) { 
+	  sleep_time = atoi(st);
+  }
+  sleep(sleep_time);
+
+  while (1) {
+
+	  file_descriptor = open(lock_path, O_RDWR | O_CREAT, 0666);
+	  lock = flock(file_descriptor, LOCK_EX | LOCK_NB);
+
+	  // If return value is 0, we have the lockfile and can continue.
+	  if (lock == 0) { return 0; }
+
+	  // Sleep for doubling lengths (1ms, 2ms, 4ms, 8ms, 16ms, etc);
+	  usleep(count * 1000);
+
+	  // Failsafe to avoid infinite looping
+	  count *= 2;
+	  if (count > 50000) { return -1; }
+  }
+
+}
+
+void return_lockfile(int file_descriptor) {
+
+  char lock_path[PATH_MAX];
+  fullpath("/lock", lock_path);
+
+  // Close the file
+  close(file_descriptor);
+
+  // Delete the lock file, as other threads check whether or not the file exists
+  remove(lock_path);
+
+}
+
 // Function to make log writes easier
 void func_log(const char * message) {
   int write_all_functions = 1;
@@ -86,7 +134,6 @@ void user_log(const char * message) {
   fputs(message, log);
   fclose(log);
 }
-
 
 /*
 helper function for db writes
@@ -215,6 +262,11 @@ ntapfuse_readlink (const char *path, char *target, size_t size)
 int
 ntapfuse_mknod (const char *path, mode_t mode, dev_t dev)
 {
+  // Get lockfile for atomic access
+  int lfd = 0; // Lock file descriptor
+  // Will return -1 on failure, otherwise continue
+  if (get_lockfile(&lfd) == -1) { return -1; }
+
   func_log("mknod called\n");
 
   uid_t user = fuse_get_context()->uid;
@@ -223,17 +275,24 @@ ntapfuse_mknod (const char *path, mode_t mode, dev_t dev)
 
   char fpath[PATH_MAX];
   fullpath (path, fpath);
+  
   int ret = mknod (fpath, mode, dev) ? -errno : 0;
   if (ret != 0)
   {
     return ret;
   }
-  return chown(fpath, user, group) ? -errno : 0;;
+  return_lockfile(&lfd);
+  return chown(fpath, user, group) ? -errno : 0;
 }
 
 int
 ntapfuse_mkdir (const char *path, mode_t mode)
 {
+  // Get lockfile for atomic access
+  int lfd = 0; // Lock file descriptor
+  // Will return -1 on failure, otherwise continue
+  if (get_lockfile(&lfd) == -1) { return -1; }
+
   func_log("mkdir called\n");
 
   uid_t user = fuse_get_context()->uid;
@@ -256,14 +315,21 @@ ntapfuse_mkdir (const char *path, mode_t mode)
     rmdir(fpath);
     return up_out;
   }
+  
+  return_lockfile(&lfd);
 
-  return chown(fpath, user, group) ? -errno : 0;;
+  return chown(fpath, user, group) ? -errno : 0;
 }
 
 //TODO: add checks for proper deletion of files!
 int
 ntapfuse_unlink (const char *path)
 {
+  // Get lockfile for atomic access
+  int lfd = 0; // Lock file descriptor
+  // Will return -1 on failure, otherwise continue
+  if (get_lockfile(&lfd) == -1) { return -1; }
+
   func_log("unlink called\n");
   char fpath[PATH_MAX];
   fullpath (path, fpath);
@@ -281,12 +347,19 @@ ntapfuse_unlink (const char *path)
     return up_out;
   }
 
+  return_lockfile(&lfd);
+
   return unlink (fpath) ? -errno : 0;
 }
 
 int
 ntapfuse_rmdir (const char *path)
 {
+  // Get lockfile for atomic access
+  int lfd = 0; // Lock file descriptor
+  // Will return -1 on failure, otherwise continue
+  if (get_lockfile(&lfd) == -1) { return -1; }
+
   func_log("rmdir called\n");
   char fpath[PATH_MAX];
   fullpath (path, fpath);
@@ -300,10 +373,13 @@ ntapfuse_rmdir (const char *path)
   uid_t user = dir_stat.st_uid;
 
   //update directory
+
   int up_out = db_update(user, -dir_stat.st_size);
   if(up_out !=0 ) {
     return up_out;
   }
+
+  return_lockfile(&lfd);
   
   return rmdir (fpath) ? -errno : 0;
 }
@@ -321,6 +397,11 @@ ntapfuse_symlink (const char *path, const char *link)
 int
 ntapfuse_rename (const char *src, const char *dst)
 {
+  // Get lockfile for atomic access
+  int lfd = 0; // Lock file descriptor
+  // Will return -1 on failure, otherwise continue
+  if (get_lockfile(&lfd) == -1) { return -1; }
+
   func_log("rename called\n");
 
   //Getting and removing the source file from user_db
@@ -329,6 +410,8 @@ ntapfuse_rename (const char *src, const char *dst)
 
   char fdst[PATH_MAX];
   fullpath (dst, fdst);
+
+  return_lockfile(&lfd);
 
   return rename (fsrc, fdst) ? -errno : 0;
 
@@ -360,6 +443,11 @@ ntapfuse_chmod (const char *path, mode_t mode)
 int
 ntapfuse_chown (const char *path, uid_t uid, gid_t gid)
 {
+  // Get lockfile for atomic access
+  int lfd = 0; // Lock file descriptor
+  // Will return -1 on failure, otherwise continue
+  if (get_lockfile(&lfd) == -1) { return -1; }
+
   func_log("chown called\n");
 
   char fpath[PATH_MAX];
@@ -381,6 +469,8 @@ ntapfuse_chown (const char *path, uid_t uid, gid_t gid)
     /* TO DO: FIND WAY TO UNDO UPDATES */
     return up_out;
   }
+  
+  return_lockfile(&lfd);
 
   return chown (fpath, uid, gid) ? -errno : 0;
 }
@@ -389,6 +479,11 @@ ntapfuse_chown (const char *path, uid_t uid, gid_t gid)
 int
 ntapfuse_truncate (const char *path, off_t length)
 {
+  // Get lockfile for atomic access
+  int lfd = 0; // Lock file descriptor
+  // Will return -1 on failure, otherwise continue
+  if (get_lockfile(&lfd) == -1) { return -1; }
+
   if (length < 0) {
     func_log("truncate called error: length < 0\n");
     return -1;
@@ -408,6 +503,8 @@ ntapfuse_truncate (const char *path, off_t length)
 
   //callstat db helper method, fails on return
   if(db_update(file_stat.st_uid, change)!=0 ) return -1;
+
+  return_lockfile(&lfd);
 
   return truncate (fpath, length) ? -errno : 0;
 }
@@ -450,6 +547,11 @@ int
 ntapfuse_write (const char *path, const char *buf, size_t size, off_t off,
 	    struct fuse_file_info *fi)
 {
+  // Get lockfile for atomic access
+  int lfd = 0; // Lock file descriptor
+  // Will return -1 on failure, otherwise continue
+  if (get_lockfile(&lfd) == -1) { return -1; }
+
   func_log("write called\n");
   // Get full path, starts out every function
   //char fpath[PATH_MAX];
@@ -460,26 +562,6 @@ ntapfuse_write (const char *path, const char *buf, size_t size, off_t off,
   // STEP 1: OPEN LOG FILE AND CHECK IF USER HAS ENOUGH AVAILABLE SPACE
   // ------------------------------------------------------------------
   // ------------------------------------------------------------------
-
-  char lock_path[PATH_MAX];
-  fullpath("/lock", lock_path);
-
-  int count = 1;
-  int lock = 0;
-  int lfd = 0;
-  while (1) {
-
-	  lfd = open(lock_path, O_RDWR | O_CREAT, 0666);
-	  lock = flock(lfd, LOCK_EX | LOCK_NB);
-	  if (lock == 0) {
-		break;
-	  }
-	  sleep(1);
-	  count += 1;
-	  if (count > 10) {
-		  return -1;
-	  }
-  }
 
   // Open db file with c standard library
   char db_path[PATH_MAX];
@@ -603,8 +685,7 @@ ntapfuse_write (const char *path, const char *buf, size_t size, off_t off,
   // Close before opening other files
   close(fi->fh);
 
-  close(lfd);
-  remove(lock_path);
+  return_lockfile(&lfd);
 
   //Return size as originally.
   return size;
